fix(test): resolve curate settle failures and backend-lifecycle unhandled rejection
Three independent test issues:

1. Curate tests: isUsageBudgetExhausted() reads ~/.claude/hud/
   .coral-cache.json via homedir() at runtime. When weekly usage >= 100%
   the scheduler silently skips curate runs, leaving isRunning() stuck.
   Fix: extract isUsageBudgetExhausted into usage-budget.ts so tests
   can vi.mock the module without production code changes. Also move
   queuedRun=false before the budget check so isRunning() accurately
   reports idle when budget is exhausted (production bug).

2. Backend-lifecycle: ensureBackend() rejection fires during
   advanceTimersByTimeAsync before the .rejects.toThrow() handler is
   attached, causing an unhandled rejection. Fix: attach the rejection
   expectation before advancing timers.
