+++ src/kb/curate/scheduler.ts
@@ -1,6 +1,3 @@
-import { readFileSync } from 'node:fs';
-import { homedir } from 'node:os';
-import { join } from 'node:path';
 import { backendLog } from '../../shared/backend-log.js';
 import { errorMessage, nowIsoString } from '../../shared/utils.js';
 import type { KbRuntime } from '../contracts.js';
@@ -65,36 +62,9 @@ export {
   type DiscoveryPromptResult,
 } from './discovery.js';
 
-const CURATE_SCHEDULE_DEBOUNCE_MS = 60 * 1000;
-const USAGE_CACHE_STALE_MS = 10 * 60 * 1000;
-const USAGE_5H_THRESHOLD = 90;
-const USAGE_WK_THRESHOLD = 100;
+import { isUsageBudgetExhausted } from './usage-budget.js';
 
-function isUsageBudgetExhausted(): boolean {
-  try {
-    const cachePath = join(homedir(), '.claude', 'hud', '.coral-cache.json');
-    const raw = JSON.parse(readFileSync(cachePath, 'utf-8')) as Record<string, unknown>;
-    const entry = raw.claude as
-      | { ts?: number; data?: { fiveHour?: number; weekly?: number }; error?: boolean }
-      | undefined;
-    if (!entry?.ts || !entry.data || entry.error) {
-      return false;
-    }
-    if (Date.now() - entry.ts > USAGE_CACHE_STALE_MS) {
-      return false;
-    }
-    const { fiveHour, weekly } = entry.data;
-    if (typeof fiveHour === 'number' && fiveHour >= USAGE_5H_THRESHOLD) {
-      return true;
-    }
-    if (typeof weekly === 'number' && weekly >= USAGE_WK_THRESHOLD) {
-      return true;
-    }
-    return false;
-  } catch {
-    return false;
-  }
-}
+const CURATE_SCHEDULE_DEBOUNCE_MS = 60 * 1000;
 
 class CurateRunError extends Error {
   readonly through: CurateCursor | null;
@@ -221,11 +191,14 @@ export function createCurateScheduler({
   }
 
   function launchQueuedRun(): void {
-    if (stopped || !runtimeStarted || activeRun !== null || !queuedRun || isUsageBudgetExhausted()) {
+    if (stopped || !runtimeStarted || activeRun !== null || !queuedRun) {
       return;
     }
 
     queuedRun = false;
+    if (isUsageBudgetExhausted()) {
+      return;
+    }
     const runController = new AbortController();
     activeRunController = runController;
     activeRun = (async () => {
